requirements.txt
faiss-cpu==1.8.0
joblib==1.4.2
ollama==0.1.9
openai==1.25.1
boto3==1.34.100
tenacity==8.2.3
numpy
pandas==2.2.2
tabulate==0.9.0
uvicorn==0.29.0
fastapi==0.111.0
pyyaml